Energy_Management/elprice/generate_plot/requirements.txt
requests
numpy
matplotlib
paho-mqtt
